notmuch-deliver: wait for the database to become unlocked
If notmuch-deliver just gives up on a locked database this can result
in mail bouncing.

This patch makes it wait a given time and then retry to open the
database. The wait time starts at a millisecond, and doubles every
time the database is still locked, until an absolute maximum time of
10 minutes is reached.